Update permit2 and narwhal (#64)
* feat: update permit2 and narwhal versions

* fix: replace package.json

* fix: lint

+++ tests/forge/SwapERC20CallParameters.t.sol
@@ -67,7 +67,7 @@ contract SwapERC20CallParametersTest is Test, Interop, DeployRouter {
 
         deal(address(USDC), from, BALANCE);
         USDC.approve(address(permit2), BALANCE);
-        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint64(block.timestamp + 1000));
+        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint48(block.timestamp + 1000));
 
         assertEq(USDC.balanceOf(from), BALANCE);
         uint256 startingRecipientBalance = RECIPIENT.balance;
@@ -98,7 +98,7 @@ contract SwapERC20CallParametersTest is Test, Interop, DeployRouter {
 
         deal(address(DAI), from, BALANCE);
         DAI.approve(address(permit2), BALANCE);
-        permit2.approve(address(DAI), address(router), uint160(BALANCE), uint64(block.timestamp + 1000));
+        permit2.approve(address(DAI), address(router), uint160(BALANCE), uint48(block.timestamp + 1000));
         assertEq(DAI.balanceOf(from), BALANCE);
         uint256 startingRecipientBalance = RECIPIENT.balance;
 
@@ -127,7 +127,7 @@ contract SwapERC20CallParametersTest is Test, Interop, DeployRouter {
 
         deal(address(USDC), from, BALANCE);
         USDC.approve(address(permit2), BALANCE);
-        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint64(block.timestamp + 1000));
+        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint48(block.timestamp + 1000));
 
         assertEq(USDC.balanceOf(from), BALANCE);
         uint256 startingRecipientBalance = RECIPIENT.balance;
@@ -155,7 +155,7 @@ contract SwapERC20CallParametersTest is Test, Interop, DeployRouter {
 
         deal(address(USDC), from, BALANCE);
         USDC.approve(address(permit2), BALANCE);
-        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint64(block.timestamp + 1000));
+        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint48(block.timestamp + 1000));
         assertEq(USDC.balanceOf(from), BALANCE);
         uint256 startingRecipientBalance = RECIPIENT.balance;
 
@@ -208,7 +208,7 @@ contract SwapERC20CallParametersTest is Test, Interop, DeployRouter {
 
         deal(address(USDC), from, BALANCE);
         USDC.approve(address(permit2), BALANCE);
-        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint64(block.timestamp + 1000));
+        permit2.approve(address(USDC), address(router), uint160(BALANCE), uint48(block.timestamp + 1000));
         assertEq(USDC.balanceOf(from), BALANCE);
         uint256 startingRecipientBalance = RECIPIENT.balance;
 
@@ -236,7 +236,7 @@ contract SwapERC20CallParametersTest is Test, Interop, DeployRouter {
         uint256 daiAmount = 2000 ether;
         deal(address(DAI), from, daiAmount);
         DAI.approve(address(permit2), daiAmount);
-        permit2.approve(address(DAI), address(router), uint160(daiAmount), uint64(block.timestamp + 1000));
+        permit2.approve(address(DAI), address(router), uint160(daiAmount), uint48(block.timestamp + 1000));
 
         assertEq(DAI.balanceOf(from), daiAmount);
         uint256 startingRecipientBalance = RECIPIENT.balance;
@@ -302,7 +302,7 @@ contract SwapERC20CallParametersTest is Test, Interop, DeployRouter {
         uint256 daiAmount = 1000 ether;
         deal(address(DAI), from, daiAmount);
         DAI.approve(address(permit2), daiAmount);
-        permit2.approve(address(DAI), address(router), uint160(daiAmount), uint64(block.timestamp + 1000));
+        permit2.approve(address(DAI), address(router), uint160(daiAmount), uint48(block.timestamp + 1000));
         assertEq(DAI.balanceOf(from), daiAmount);
         uint256 startingRecipientBalance = RECIPIENT.balance;
